Dal: refactor to interface and its implementation
